Add stock Audit Create miw-upm/betca-tpv#624

+++ src/main/java/es/upm/miw/betca_tpv_core/infrastructure/mongodb/daos/DatabaseSeederDev.java
@@ -446,7 +446,7 @@ public class DatabaseSeederDev {
         stockAudit.setCreationDate(creationDate);
         stockAudit.setCloseDate(creationDate.plusDays(15));
         stockAudit.setArticlesWithoutAudit(articles);
-        stockAudit.setLossValue(numArticles * 10);
+        stockAudit.setLossValue(BigDecimal.valueOf(numArticles * 10L));
         stockAudit.setLosses(losses);
 
         return stockAudit;
